Add files via upload
This is just a run down of how to pull information from any HTML site with tables; Wikipedia used as an example. Once you pull it you need to scrub the data to get what you want to write to file for analysis later, so that is included as well.